// frontend/src/constants/enum.ts
export enum TaskStatus {
  PENDING = 'Pending',
  IN_PROGRESS = 'In Progress',
  COMPLETED = 'Completed',
  IN_REVIEW = 'In Review',
  UNASSIGNED = 'Unassigned'
}

export enum TaskPriority {
  LESS_IMPORTANT = 'Less Important',
  IMPORTANT = 'Important',
  VERY_IMPORTANT = 'Very Important'
}

export enum Mode {
  LIGHT = 'light',
  DARK = 'dark'
}